HDX-9781 Move national_risk view as table
