+++ src/hapi_schema/db_national_risk.py
@@ -9,7 +9,6 @@ from sqlalchemy import (
     Float,
     ForeignKey,
     Integer,
-    String,
     select,
 )
 from sqlalchemy.orm import Mapped, mapped_column, relationship
@@ -93,25 +92,3 @@ view_params_national_risk = ViewParams(
         )
     ),
 )
-
-
-class DBNationalRiskVAT(Base):
-    __tablename__ = "national_risk_vat"
-    resource_hdx_id: Mapped[str] = mapped_column(String(36))
-    location_ref: Mapped[int] = mapped_column(Integer, primary_key=True)
-    risk_class: Mapped[str] = mapped_column(String(9))
-    global_rank: Mapped[int] = mapped_column(Integer)
-    overall_risk: Mapped[float] = mapped_column(Float)
-    hazard_exposure_risk: Mapped[float] = mapped_column(Float)
-    vulnerability_risk: Mapped[float] = mapped_column(Float)
-    coping_capacity_risk: Mapped[float] = mapped_column(Float)
-    meta_missing_indicators_pct: Mapped[float] = mapped_column(Float)
-    meta_avg_recentness_years: Mapped[float] = mapped_column(Float)
-    reference_period_start: Mapped[datetime] = mapped_column(
-        DateTime, primary_key=True
-    )
-    reference_period_end: Mapped[datetime] = mapped_column(
-        DateTime, primary_key=True
-    )
-    location_code: Mapped[str] = mapped_column(String(128))
-    location_name: Mapped[str] = mapped_column(String(512))
