+++ src/hapi_schema/db_views_as_tables.py
@@ -151,3 +151,29 @@ class DBLocationVAT(Base):
     reference_period_end: Mapped[datetime] = mapped_column(
         DateTime, index=True
     )
+
+
+class DBNationalRiskVAT(Base):
+    __tablename__ = "national_risk_vat"
+    resource_hdx_id: Mapped[str] = mapped_column(String(36))
+    location_ref: Mapped[int] = mapped_column(Integer, primary_key=True)
+    risk_class: Mapped[str] = mapped_column(String(9))
+    global_rank: Mapped[int] = mapped_column(Integer)
+    overall_risk: Mapped[float] = mapped_column(Float)
+    hazard_exposure_risk: Mapped[float] = mapped_column(Float)
+    vulnerability_risk: Mapped[float] = mapped_column(Float)
+    coping_capacity_risk: Mapped[float] = mapped_column(Float)
+    meta_missing_indicators_pct: Mapped[float] = mapped_column(
+        Float, nullable=True
+    )
+    meta_avg_recentness_years: Mapped[float] = mapped_column(
+        Float, nullable=True
+    )
+    reference_period_start: Mapped[datetime] = mapped_column(
+        DateTime, primary_key=True
+    )
+    reference_period_end: Mapped[datetime] = mapped_column(
+        DateTime, index=True
+    )
+    location_code: Mapped[str] = mapped_column(String(128), index=True)
+    location_name: Mapped[str] = mapped_column(String(512), index=True)
